Adding create command from a user call of /sbin/mkfs.ext2.
Now, filesystems are created on the loop device mounted as 'none'. It is way easier than creating from a Volume Handle, but I don't like it. mounting and unmounting are expensive operations, not needed in this case. I did not know it was so hard to just initialize a simple filesystem as a user!