Adding a ForkableCore to the Qt UI to fix a bug.
Notes:
The library GostCrypt uses fuse_main. fuse_main uses fork to run in background.
Qt does not run well with fork(), so I had to create this 'little' workaround, where the Qt interfaces execv itself as a command-line application that can be forked. Any application using GostCrypt who can not be forked may use the ForkableCore files to be able to mount and create volumes. Else, the command-line can be used too.